Automatically fixed clippy warnings

+++ src/verbose/values/array_i64.rs
@@ -67,9 +67,9 @@ impl<'a> ArrayI64<'a> {
                 buf.try_extend_from_slice(var_info.name.as_bytes())?;
                 if buf.remaining_capacity() > var_info.unit.len() + 2 {
                     // Safe as capacity is checked earlier
-                    unsafe { buf.push_unchecked(u8::from(0)) };
+                    unsafe { buf.push_unchecked(0) };
                     let _ = buf.try_extend_from_slice(var_info.unit.as_bytes());
-                    unsafe { buf.push_unchecked(u8::from(0)) };
+                    unsafe { buf.push_unchecked(0) };
                 } else {
                     return Err(CapacityError::new(()));
                 }
@@ -87,9 +87,9 @@ impl<'a> ArrayI64<'a> {
                 buf.try_extend_from_slice(var_info.name.as_bytes())?;
                 if buf.remaining_capacity() > var_info.unit.len() + 2 {
                     // Safe as capacity is checked earlier
-                    unsafe { buf.push_unchecked(u8::from(0)) };
+                    unsafe { buf.push_unchecked(0) };
                     let _ = buf.try_extend_from_slice(var_info.unit.as_bytes());
-                    unsafe { buf.push_unchecked(u8::from(0)) };
+                    unsafe { buf.push_unchecked(0) };
                 } else {
                     return Err(CapacityError::new(()));
                 }
